requirements.txt
# Flask core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0

# Serialization
marshmallow==3.21.1
flask-marshmallow==1.2.1

# Async & API requests
aiohttp==3.10.5

# Telegram Bot
aiogram==3.13.1

# Utils
python-dotenv==1.0.1
requests==2.32.3

# Optional (for scheduling background checks)
APScheduler==3.10.4